LayoutTests/platform/ios-sim-deprecated/iphone/limits/no-image-limit-expected.txt

CONSOLE MESSAGE: line 15: TypeError: HTMLParagraphElement is not a function (evaluating 'description')
